Move Callbacks interface to API again
