+++ api/src/main/java/mb/pie/api/Callbacks.java
@@ -1,7 +1,5 @@
-package mb.pie.runtime;
+package mb.pie.api;
 
-import mb.pie.api.Task;
-import mb.pie.api.TaskKey;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.io.Serializable;
