+++ runtime/src/main/java/mb/pie/runtime/MapCallbacks.java
@@ -1,5 +1,6 @@
 package mb.pie.runtime;
 
+import mb.pie.api.Callbacks;
 import mb.pie.api.TaskKey;
 import org.checkerframework.checker.nullness.qual.Nullable;
