+++ runtime/src/main/java/mb/pie/runtime/exec/BottomUpRunner.java
@@ -19,7 +19,7 @@ import mb.pie.api.TaskKey;
 import mb.pie.api.TaskRequireDep;
 import mb.pie.api.exec.CancelToken;
 import mb.pie.api.exec.ExecReason;
-import mb.pie.runtime.Callbacks;
+import mb.pie.api.Callbacks;
 import mb.resource.ResourceKey;
 import mb.resource.ResourceService;
 import org.checkerframework.checker.nullness.qual.Nullable;
